@@ -1,12 +1,8 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import App from './components/app/App';
-import './style/style.scss';
+import React, { StrictMode } from "react";
+import ReactDOM from "react-dom/client";
+import App from "./components/app/App";
+import "./style/style.scss";
 
-ReactDOM.render(
-  <React.StrictMode>
+ReactDOM.createRoot(document.getElementById("root")).render(
     <App />
-  </React.StrictMode>,
-  document.getElementById('root')
 );
-
